Terminology did:btcr2 Updates
Updates references to did:btc1 to did:btcr2, including:
- BTC1 Beacon to BTCR2 Beacon
- BTC1 Update Announcement to BTCR2 Update Announcement
- BTC1 Update to BTCR2 Update
- Unsecured BTC1 Update to Unsecured BTCR2 Update
- Pending BTC1 Update to Pending BTCR2 Update
- Announced BTC1 Update to Announced BTCR2 Update

Note: Given the scope of the name change, I've left invisible link redirects in place for the BTC1 versions of the terms above so they will link to the correct BTCR2 term in the Terminology Section, just in case we miss one in this round of edits. These links are visible when editing the chapter or viewing code, but not in the published spec, and can be removed if unnecessary.