Add terminology and algorithm reference links
There is a top-level `includes` directory that all top-level chapters
can use to include common links. The `includes.md` file is a shortcut to
include all of the inclusion files.

Relative links are a pain. The `crud` subsections have a duplicate of
the `includes` directory. The only difference compared to the top-level
is that the subsection links need the `../` parent directory notation
for relative links to work correctly.

A preprocessor plugin might be a better way to manage global/relative
links. The duplication is going to be problematic at some point!