Integrate pagetoc into the normal TOC sidebar
The terminology and algorithms chapters are the longest. They make the
best test cases for headers getting mapped into the chapters list.

The mdbook-pagetoc preprocessor uses JavaScript to insert the header
links. I don't know why. So, ok, it injects the TOC on the client
side. Fine. But at least use normal HTML lists and CSS to represent
the structure!

The CSS has been completely replaced with just a simple "dimming" filter.
Otherwise uses the existing sidebar CSS.

It is also possible to make the headers collapsible in the TOC. (This is
done by `[output.html.fold]`.) But I haven't really explored that option
yet.