Don't forget that errors are a thing!
Maybe there is a better way to write this loop with its various exits.
It is easier to describe an algorithm when it has a single exit. E.g.,
no early exits. But that rules out raising errors and handling
`versionTime`!

I don't really know what to do about it.