Show all tab contents when printing
When printing the specification, examples hidden by tabs would not be
printed. This fixes the issue by forcing all tab contents to show
themselves for printers.

The tabs bar is hidden for printers because it is unnecessary UI for
a static rendering. Instead, labels are added to tab section contents
for printers (hidden on screens). Labels are also hidden from printers
if the rest of the section content is empty.

Page breaks are also properly inserted for printers around the tab
section, _not_ between the section's label and its sibling example.

CSS media queries to the rescue!