Various fixes for the Resolve chapter.
Fixes block-confirmations and sorting updates by block-height after
targetVersionId. This allows `didResolutionMetadata.confirmations` to be
returned to the caller correctly.

Fixes a subtle error in "Process SMT Beacons" that was missing a point
about how `smt_root` is used.

Added pseudocode blocks that are more formally defined than the prose.

Removed uses of imperative "passing arguments" language by in-lining the
"remove `update.proof`" step into "Check Duplicate Update". I think of
the chapter as a class and the sub-sections as methods. All sub-sections
have implicit access to class state instantiated at the top (e.g.
`this`).

There is only one remaining use of "pass" and it is for calling out to
an algorithm defined in another specification.

Finally, removed all of the old brainstorming information at the end of
the chapter.

Only a few TODOs left. Nice!