More improvements to the Resolve chapter
Corrects a bug where `unsigned_update_history` was not updated.

Put all pseudocode behind tabs, hidden by default.

Clean up the Resolve operation outputs: Don't describe what these mean,
just tell what to put there.

`block_confirmations` was added as internal state to make
didResolutionMetadata.confirmations work.

Use "let" as a means of _introducing_ an identifier. Use "assign",
"push", "pop", etc. as a means of _updating_ an existing identifier.

Use active voice by consistently putting conditions at the end of the
statement:

"An error MUST be raised if the condition is met."

INSTEAD OF:

"If the condition is met, an error MUST be raised."

OR EVEN:

"If the condition is met, MUST raise an error."