Fix Deactivate formatting and bug in Resolve
Cool! Reading the tiny Deactivate chapter tipped me off to a bug in
Resolve! We did not have the stated condition that requires
`deactivated` to be treated by the resolver as permanent.

Now we do.

And even more revealing is that this condition is quite old but has
never been described in the resolution algorithm, from what I can see.

Here is the condition in the initial commit:

https://github.com/dcdpr/did-btcr2/blob/bb29163a7fb144421e676a266e999d4e5eeb2e5a/README.md?plain=1#L596

This bug has always existed.