Remove unused references
Also fixes some incorrect citations.

For instance, "... according to {{#cite FOO}}" reads oddly. Because the
"FOO" citation identifier can be (and often is) numeric.

"... according to [9]" would be quite confusing if reading the spec on
paper. Though you could probably skip to the last page and figure out
what it is trying to tell you, the better option is to just say the
right thing without relying on the citation identifier being meaningful.