Additions for the Update chapter
Splits update creation into "Unsigned" and "Signed" sub-sections.

Adds missing assertions required when signing the update.

Adds a new "Data Integrity Config" data structure which is defined as a
Data Integrity Proof with the `proofValue` property omitted. The data
structure is created with a template and passed to the
`cryptosuite.createProof` method to produce `update.proof`.

Adds clarity around SHA-256 hashes described in the Data Structures
chapter.

Fixes the encoding used on public keys and signatures. (The encoding is
required by `BIP340-Cryptosuite` but is specified by `CID`.)

Removes Algos that are now covered:

* Algo 15: I guess this just shows an example of updating a DID
  document. There is no reason to specify an example as an algorithm.
* Algo 17: Covered in Updates chapter.

Remaining Algos are for update announcements (which are TBD).